requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-writable-nested==0.7.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.1
uritemplate==4.1.1
